HDDS-2609. Move plain unit tests out of integration tests
