+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestKeyPurging.java
@@ -28,6 +28,7 @@ import org.apache.hadoop.ozone.client.OzoneClientFactory;
 import org.apache.hadoop.ozone.client.OzoneVolume;
 import org.apache.hadoop.ozone.client.io.OzoneOutputStream;
 import org.apache.hadoop.ozone.container.ContainerTestHelper;
+import org.apache.hadoop.ozone.container.TestHelper;
 import org.apache.hadoop.test.GenericTestUtils;
 import org.junit.After;
 import org.junit.Assert;
@@ -101,7 +102,7 @@ public class TestKeyPurging {
     for (int i = 1; i <= NUM_KEYS; i++) {
       String keyName = keyBase + "-" + i;
       keys.add(keyName);
-      OzoneOutputStream keyStream = ContainerTestHelper.createKey(
+      OzoneOutputStream keyStream = TestHelper.createKey(
           keyName, ReplicationType.STAND_ALONE, ReplicationFactor.ONE,
           KEY_SIZE, store, volumeName, bucketName);
       keyStream.write(data);
