+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/ratis/TestOMRatisSnapshotInfo.java
@@ -15,9 +15,8 @@
  * the License.
  */
 
-package org.apache.hadoop.ozone.om.snapshot;
+package org.apache.hadoop.ozone.om.ratis;
 
-import org.apache.hadoop.ozone.om.ratis.OMRatisSnapshotInfo;
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
